cloudfunctions/requirements.txt
flask==1.1.2
numpy==1.18.5
google-cloud-storage